google-fit-pipeline/pipeline/assets/ingestion/requirements.txt
# TODO: Add Python dependencies required by the python ingestion asset.
# - Keep this intentionally minimal.
# - Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
# Example:
# pandas==2.2.0
# requests==2.31.0

google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-api-python-client==2.193.0
pandas==2.2.3